==> Scripts/1.0_WEIGHT_LENGTH.Rmd <==
---
title: "Weight and Length"
author: "Hamza"
date: "23/11/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

### Load packages
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#checks for installation and loads packages
pacman::p_load(lmerTest,ggThemeAssist,rptR,lme4,readxl, tidyr, dplyr, magrittr, lubridate, stringr, purrr,
               sjPlot,ggplot2,lubridate,wesanderson,ggbeeswarm,emmeans,patchwork,viridis,nlme,Rmisc,ggpubr,
               stargazer, kableExtra)

# Load custom functions
source("functions_learning.R")
source("functions_repeatability.R")

#Run this after all the coding has been run with 10,000 bootstraps
#save.image(file='myEnvironment.RData') 
#load('myEnvironment.RData')
```

#Import and filter data
```{r, include = FALSE}
#Master file with raw data
Weight_Length_Data <- read.csv("./Data/Weight_Length/Weight_Length.csv")

str(Weight_Length_Data)

Weight_Length_Data$Age_Weeks <- as.numeric(Weight_Length_Data$Age_Weeks)
# Weight_Length_Data$Age_Weeks <- as.numeric(as.factor(Weight_Length_Data$Age_Weeks))



#Subseting data by control and treatment
Weight_Length_Maternal <- subset(Weight_Length_Data, Weight_Length_Data$Set == "Maternal") 
Weight_Length_Paternal <- subset(Weight_Length_Data, Weight_Length_Data$Set == "Paternal") 
Weight_Length_Control <- subset(Weight_Length_Data, Weight_Length_Data$Set == "Within Control") 
Weight_Length_Treatment <- subset(Weight_Length_Data, Weight_Length_Data$Set == "Within Treatment") 

Weight_Length_Age30 <- subset(Weight_Length_Data, Weight_Length_Data$Age_Weeks == "30") 
```

### REPEATABILITY ANALYSIS FOR BODY WEIGHT AND LENGTH 
1) CHECKING NORMALITY ASSUMPTIONS WITH MIXED MODELS AND A HISTOGRAM OF RESIDUALS
2) PERFORMING REPEATABILITY ANALYSIS AND EXTRACTING WITHIN-AND BETWEEN-INDIVIDUAL VARIANCES AFTER A Z TRANSFORMATION (FOR BOTH CONTROL AND TREATMENT)
3) CALCULATING DIFFERENCES BETWEEN REPEATABILITIES
4) FINAL MIXED MODEL USING LME FUNCTION
5) CALCULATING VARIANCE DIFFERENCES BETWEEN CONTROL AND TREATMENT GROUPS

# Body Weight
```{r, echo=FALSE}

#Check Normality with mixed model and then histogram of residuals

Model_Body_Weight <- lmer(Weight_g ~ Set + Sex + Age_Weeks + (1 | Fish_ID), data = Weight_Length_Data)
summary(Model_Body_Weight)
tab_model(Model_Body_Weight)
hist(residuals(Model_Body_Weight))

AIC(Model_Body_Weight)

emmeans_weight <- emmeans(Model_Body_Weight, ~c(Set))
pairs(emmeans_weight)

#Calculating repeatabilities (with week) 
# Will run with 10,000 bootstraps if required but not required atm

rpt_weight2(Weight_Length_Control) -> rpt_body_weight_control 
rpt_body_weight_control
rpt_weight2(Weight_Length_Treatment) -> rpt_body_weight_treatment 
rpt_body_weight_treatment

rpt_weight2(Weight_Length_Maternal) -> rpt_body_weight_maternal 
rpt_body_weight_maternal
rpt_weight2(Weight_Length_Paternal) -> rpt_body_weight_paternal
rpt_body_weight_paternal


# Calculating difference in variance between control and treatment group 
Model_Weight_1 <- lme(Weight_g ~ Set + Sex + scale(Age_Weeks, scale = FALSE), random = ~ 1| Fish_ID, data = Weight_Length_Data) #Mixed model without variance structure
summary(Model_Weight_1)
tab_model(Model_Weight_1)
Model_Weight_2 <- lme(Weight_g ~ Set + Sex + scale(Age_Weeks, scale = FALSE), random = ~ 1| Fish_ID, weights = varIdent(form=~1|Set), data = Weight_Length_Data) #mixed model with variance structure varIdent
summary(Model_Weight_2)
anova(Model_Weight_1, Model_Weight_2) #calculating difference between two models to find difference in variance
```


#Plot
```{r}
Weight_Length_Data_reorder <- Weight_Length_Data %>%
  mutate(Set = fct_relevel(Set, 
            "Within Control", "Maternal", "Paternal", 
            "Within Treatment"))

violin_plot_bw <- function(df,param,lab1,lab2) { #only need to add dataframe, parameter, and labels for axes
  p <- ggplot(data = df,aes(x = Sex, y = param, fill = Set))+
      scale_fill_manual(values=c("#009E73","#E69F00","#56B4E9","#999999"), labels = c("F1C", "F1M", "F1P", "F1T"))+ #color scheme
    geom_violin(alpha=0.4, position = position_dodge(width = .75),size=0.5,color="black")+ #violin plot to display distribution density
    geom_boxplot(notch = TRUE,  outlier.size = -1, color="black",lwd=0.9, alpha = 0.5)+#boxplot to display quantiles
    geom_point( shape = 21,size=2, position = position_jitterdodge(), color="black",alpha=1)+#scatterplot of individual data points
    theme_pubr(base_family = "")+
    labs(title=lab1,y = lab2)+#labels
    rremove("legend.title")+
    theme(panel.border = element_rect(colour = "black", fill=NA, size=1),
    axis.ticks = element_line(size=2,color="black"),axis.ticks.length=unit(0.2,"cm"),
    legend.position = c(0.92, 0.85))+
    font("xylab",size=13)+font("xy",size=13)+
    font("xy.text", size = 13)+
    font("legend.text",size = 13)+
    theme(legend.position = "right")+
    theme(axis.title.x=element_blank())+
    theme(text=element_text(family="Times New Roman", size=12))
  return(p)
  
}

violin_plot_bw(Weight_Length_Data_reorder, Weight_Length_Data_reorder$Weight_g, "Body weight (g) by age (weeks)", "g")+
  facet_grid(~Age_Weeks)







```
