==> Scripts/3.1_ANXIETY_SEX.Rmd <==
---
title: "Anxiety Sex Analysis F0"
author: "Hamza"
date: "23/10/2020"
output: html_document
---

### Load packages
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#checks for installation and loads packages
pacman::p_load(lmerTest,ggThemeAssist,rptR,lme4,readxl, tidyr, dplyr, magrittr, lubridate, stringr, purrr,
               sjPlot,ggplot2,lubridate,wesanderson,ggbeeswarm,emmeans,patchwork,viridis,nlme,Rmisc,ggpubr,
               stargazer)

# Load custom functions
source("../Scripts/functions_repeatability.R") #load custom functions
```

#Subset data by sex
```{r}
#Subset by sex
Anxiety_Control_Male <- subset(Anxiety_Control, Anxiety_Control$Sex == "male")
Anxiety_Control_Female <- subset(Anxiety_Control, Anxiety_Control$Sex == "female")
Anxiety_Treatment_Male <- subset(Anxiety_Treatment, Anxiety_Treatment$Sex == "male")
Anxiety_Treatment_Female <- subset(Anxiety_Treatment, Anxiety_Treatment$Sex == "female")
Anxiety_Maternal_Male <- subset(Anxiety_Maternal, Anxiety_Maternal$Sex == "male")
Anxiety_Maternal_Female <- subset(Anxiety_Maternal, Anxiety_Maternal$Sex == "female")
Anxiety_Paternal_Male <- subset(Anxiety_Paternal, Anxiety_Paternal$Sex == "male")
Anxiety_Paternal_Female <- subset(Anxiety_Paternal, Anxiety_Paternal$Sex == "female")

```

### MAIN ANALYSIS FOR ALL BEHAVIORAL PARAMETERS BY SEX AND TANK
1) CHECKING NORMALITY ASSUMPTIONS WITH MIXED MODELS AND A HISTOGRAM OF RESIDUALS
2) PERFORMING REPEATABILITY ANALYSIS FOR EACH SEX IN EACH TANK SEPARATELY (USING SUBSETS) AND THEN WITH BOTH DATA COMBINED FOR EACH TANK
3) CALCULATING DIFFERENCES BETWEEN MALE AND FEMALE REPEATABILITIES IN EACH TANK

THESE STEPS ARE REPEATED FOR ALL BEHAVIORAL PARAMETERS, REFER TO THE FIRST BEHAVIORAL PARAMETER FOR COMMENTS AND REFERENCE 

### Total distance travelled
```{r}

#Repeatability analysis males
rpt_tot_dist(Anxiety_Control_Male) -> Control_Tot_Dist_Male 
Control_Tot_Dist_Male
rpt_tot_dist(Anxiety_Treatment_Male) -> Treatment_Tot_Dist_Male 
Treatment_Tot_Dist_Male

#Repeatability analysis females
rpt_tot_dist(Anxiety_Control_Female) -> Control_Tot_Dist_Female 
Control_Tot_Dist_Female
rpt_tot_dist(Anxiety_Treatment_Female) -> Treatment_Tot_Dist_Female
Treatment_Tot_Dist_Female

#Repeatability analysis males
rpt_tot_dist(Anxiety_Maternal_Male) -> Maternal_Tot_Dist_Male 
Maternal_Tot_Dist_Male
rpt_tot_dist(Anxiety_Paternal_Male) -> Paternal_Tot_Dist_Male 
Paternal_Tot_Dist_Male

#Repeatability analysis females
rpt_tot_dist(Anxiety_Maternal_Female) -> Maternal_Tot_Dist_Female 
Maternal_Tot_Dist_Female
rpt_tot_dist(Anxiety_Paternal_Female) -> Paternal_Tot_Dist_Female
Paternal_Tot_Dist_Female


```


### Time spent in low zone
```{r}

#Repeatability analysis males
rpt_low_dur(Anxiety_Control_Male) -> Control_low_dur_Male 
Control_low_dur_Male
rpt_low_dur(Anxiety_Treatment_Male) -> Treatment_low_dur_Male 
Treatment_low_dur_Male

#Repeatability analysis females
rpt_low_dur(Anxiety_Control_Female) -> Control_low_dur_Female 
Control_low_dur_Female
rpt_low_dur(Anxiety_Treatment_Female) -> Treatment_low_dur_Female
Treatment_low_dur_Female

#Repeatability analysis males
rpt_low_dur(Anxiety_Maternal_Male) -> Maternal_low_dur_Male 
Maternal_low_dur_Male
rpt_low_dur(Anxiety_Paternal_Male) -> Paternal_low_dur_Male 
Paternal_low_dur_Male

#Repeatability analysis females
rpt_low_dur(Anxiety_Maternal_Female) -> Maternal_low_dur_Female 
Maternal_low_dur_Female
rpt_low_dur(Anxiety_Paternal_Female) -> Paternal_low_dur_Female
Paternal_low_dur_Female

```

### Time spent in mid zone
```{r}

#Repeatability analysis males
rpt_mid_dur(Anxiety_Control_Male) -> Control_mid_dur_Male 
Control_mid_dur_Male
rpt_mid_dur(Anxiety_Treatment_Male) -> Treatment_mid_dur_Male 
Treatment_mid_dur_Male

#Repeatability analysis females
rpt_mid_dur(Anxiety_Control_Female) -> Control_mid_dur_Female 
Control_mid_dur_Female
rpt_mid_dur(Anxiety_Treatment_Female) -> Treatment_mid_dur_Female
Treatment_mid_dur_Female

#Repeatability analysis males
rpt_mid_dur(Anxiety_Maternal_Male) -> Maternal_mid_dur_Male 
Maternal_mid_dur_Male
rpt_mid_dur(Anxiety_Paternal_Male) -> Paternal_mid_dur_Male 
Paternal_mid_dur_Male

#Repeatability analysis females
rpt_mid_dur(Anxiety_Maternal_Female) -> Maternal_mid_dur_Female 
Maternal_mid_dur_Female
rpt_mid_dur(Anxiety_Paternal_Female) -> Paternal_mid_dur_Female
Paternal_mid_dur_Female

```


### Time spent in high zone
```{r}

#Repeatability analysis males
rpt_high_dur(Anxiety_Control_Male) -> Control_high_dur_Male 
Control_high_dur_Male
rpt_high_dur(Anxiety_Treatment_Male) -> Treatment_high_dur_Male 
Treatment_high_dur_Male

#Repeatability analysis females
rpt_high_dur(Anxiety_Control_Female) -> Control_high_dur_Female 
Control_high_dur_Female
rpt_high_dur(Anxiety_Treatment_Female) -> Treatment_high_dur_Female
Treatment_high_dur_Female

#Repeatability analysis males
rpt_high_dur(Anxiety_Maternal_Male) -> Maternal_high_dur_Male 
Maternal_high_dur_Male
rpt_high_dur(Anxiety_Paternal_Male) -> Paternal_high_dur_Male 
Paternal_high_dur_Male

#Repeatability analysis females
rpt_high_dur(Anxiety_Maternal_Female) -> Maternal_high_dur_Female 
Maternal_high_dur_Female
rpt_high_dur(Anxiety_Paternal_Female) -> Paternal_high_dur_Female
Paternal_high_dur_Female

```


### Time spent freezing
```{r}

#Repeatability analysis males
rpt_freezing_dur(Anxiety_Control_Male) -> Control_freezing_dur_Male 
Control_freezing_dur_Male
rpt_freezing_dur(Anxiety_Treatment_Male) -> Treatment_freezing_dur_Male 
Treatment_freezing_dur_Male

#Repeatability analysis females
rpt_freezing_dur(Anxiety_Control_Female) -> Control_freezing_dur_Female 
Control_freezing_dur_Female
rpt_freezing_dur(Anxiety_Treatment_Female) -> Treatment_freezing_dur_Female
Treatment_freezing_dur_Female

#Repeatability analysis males
rpt_freezing_dur(Anxiety_Maternal_Male) -> Maternal_freezing_dur_Male 
Maternal_freezing_dur_Male
rpt_freezing_dur(Anxiety_Paternal_Male) -> Paternal_freezing_dur_Male 
Paternal_freezing_dur_Male

#Repeatability analysis females
rpt_freezing_dur(Anxiety_Maternal_Female) -> Maternal_freezing_dur_Female 
Maternal_freezing_dur_Female
rpt_freezing_dur(Anxiety_Paternal_Female) -> Paternal_freezing_dur_Female
Paternal_freezing_dur_Female

```


### Latency to high
```{r}
#Repeatability analysis males
rpt_latency(Anxiety_Control_Male) -> Control_latency_Male 
Control_latency_Male
rpt_latency(Anxiety_Treatment_Male) -> Treatment_latency_Male 
Treatment_latency_Male

#Repeatability analysis females
rpt_latency(Anxiety_Control_Female) -> Control_latency_Female 
Control_latency_high_Female
rpt_latency(Anxiety_Treatment_Female) -> Treatment_latency_Female
Treatment_latency_high_Female

#Repeatability analysis males
rpt_latency(Anxiety_Maternal_Male) -> Maternal_latency_Male 
Maternal_latency_Male
rpt_latency(Anxiety_Paternal_Male) -> Paternal_latency_Male 
Paternal_latency_Male

#Repeatability analysis females
rpt_latency(Anxiety_Maternal_Female) -> Maternal_latency_Female 
Maternal_latency_Female
rpt_latency(Anxiety_Paternal_Female) -> Paternal_latency_Female
Paternal_latency_Female

```

### Entries to high zone
```{r}

#Repeatability analysis males
rpt_freq(Anxiety_Control_Male) -> Control_freq_Male 
Control_freq_Male
rpt_freq(Anxiety_Treatment_Male) -> Treatment_freq_Male 
Treatment_freq_high_Male

#Repeatability analysis females
rpt_freq(Anxiety_Control_Female) -> Control_freq_Female 
Control_freq_Female
rpt_freq(Anxiety_Treatment_Female) -> Treatment_freq_Female
Treatment_freq_Female

#Repeatability analysis males
rpt_freq(Anxiety_Maternal_Male) -> Maternal_freq_Male 
Maternal_freq_Male
rpt_freq(Anxiety_Paternal_Male) -> Paternal_freq_Male 
Paternal_freq_Male

#Repeatability analysis females
rpt_freq(Anxiety_Maternal_Female) -> Maternal_freq_Female 
Maternal_freq_Female
rpt_freq(Anxiety_Paternal_Female) -> Paternal_freq_Female
Paternal_freq_Female

```
