==> Scripts/6.0_IMPORT_DATA_LEARNING.Rmd <==
---
title: "01_import_data"
output: html_document
---

### Load packages
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#Package conflict causing issue with loading data, update which packages need to be loaded
#checks for installation and loads packages
pacman::p_load(tidyverse, sjPlot, lme4, lmerTest, nlme, ggplot2, rptR, emmeans, wesanderson)

# Load custom function to extract data 
source("../Scripts/functions_learning.R")
source("../Scripts/functions_repeatability.R")

#load('myEnvironment.RData')
```


```{r, warning=FALSE}
## importing csv's & transforming data
path <- paste0("../Data/Learning/Day", 1:16, "/")

all_path <- map(path,  ~list.files(path = .)[str_detect(list.files(path = .), ".csv")] ) %>% map2(.,path,~ paste0(.y, .x))

dat <- map(unlist(all_path), import_csv) %>% bind_rows() %>% as_tibble() %>% group_by(filepath, date, marking, type, test, unit, variable, arena) %>% summarise(total = sum(number), mean = mean(number)) %>% filter(variable == "ZONE_TIMERS") %>% group_by(type, marking, test, unit, filepath, date) %>% summarise(sum_time = sum(total), arena = unique(arena)) %>% arrange(filepath, marking, type, test) %>% mutate(time_min = if_else(type == "BASELINE", sum_time/5, sum_time/1)) %>% add_day %>% add_exp %>% add_zone %>% add_tod %>% add_set %>% mutate(day = as.numeric(day)) %>% rename(fishID = marking)

dat
```

```{r}
dat %>% filter(unit %in% c("UNIT1", "UNIT2", "UNIT3", "UNIT4", "UNIT5", "UNIT6", "PROBE1", "PROBE2")) %>%
  group_by(type, fishID, test, filepath, date) %>%
  summarise(time = mean(time_min), arena = unique(arena)) %>%
  add_diff -> dat_2probe #We will use this dataset as it takes into acocunt the first 2 mins of the probe period, which the literature says is the most important as fish tend to forget after 2 mins, however we have options below for analysis also

dat %>% filter(unit %in% c("UNIT5", "UNIT6", "PROBE1", "PROBE2")) %>%
  group_by(type, fishID, test, filepath, day) %>%
  summarise(time = mean(time_min), arena = unique(arena)) %>%
  add_diff -> dat_2base

dat %>% filter(unit %in% c("UNIT1", "UNIT2", "UNIT3", "UNIT4", "UNIT5", "UNIT6", "PROBE1", "PROBE2","PROBE3", "PROBE4", "PROBE5")) %>%
  group_by(type, fishID, test, filepath, day) %>%
  summarise(time = mean(time_min), arena = unique(arena)) %>%
  add_diff -> dat_5probe

dat %>% filter(unit %in% c("UNIT1", "UNIT2", "UNIT3", "UNIT4", "UNIT5", "UNIT6", "PROBE1")) %>%
  group_by(type, fishID, test, filepath, day) %>%
  summarise(time = mean(time_min), arena = unique(arena)) %>%
  add_diff -> dat_1probe

sex <- read.csv("../Data/Learning/ID_Sex_Tank.csv") %>% rename(fishID = Fish_ID) %>% rename(sex = Sex)

left_join(sex, dat_2probe, by = "fishID") -> p2
left_join(sex, dat_2base, by = "fishID") -> b2
left_join(sex, dat_5probe, by = "fishID") -> p5
left_join(sex, dat_1probe, by = "fishID") -> p1


p2 %>% add_set %>% add_tod %>% filter(test == "BLUCS") -> p2cs
b2%>% add_set %>% add_tod %>% filter(test == "BLUCS") -> b2cs
p5 %>% add_set %>% add_tod %>% filter(test == "BLUCS") -> p5cs
p1 %>% add_set %>% add_tod %>% filter(test == "BLUCS") -> p1cs
```
