==> Scripts/6.1_LEARNING_ANALYSIS.Rmd <==
---
title: "analysis"
output: html_document
---

### Load packages
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#checks for installation and loads packages
pacman::p_load(tidyverse, sjPlot, lme4, lmerTest, nlme, ggplot2, rptR, emmeans, wesanderson, ggpubr, extrafonts, stargazer, patchwork)         

# Load custom function to extract data 
source("../Scripts/functions_learning.R")
```

#Subset data into control and treatment groups and then by sex
```{r}
#Subsetting data by control and treatment
Learning_Control <- subset(p2cs, p2cs$Tank == "Within_Control") 
Learning_Treatment <- subset(p2cs, p2cs$Tank == "Within_Treatment") 
Learning_Maternal <- subset(p2cs, p2cs$Tank == "Maternal") 
Learning_Paternal <- subset(p2cs, p2cs$Tank == "Paternal") 

#Loading rpt models
load("./Models/rpt_learning_control.rdata")
load("./Models/rpt_learning_treatment.rdata")
load("./Models/rpt_learning_Maternal.rdata")
load("./Models/rpt_learning_Paternal.rdata")
```

### MAIN ANALYSIS 
1) CHECKING NORMALITY ASSUMPTIONS WITH MIXED MODELS AND A HISTOGRAM OF RESIDUALS
2) PERFORMING REPEATABILITY ANALYSIS FOR EACH GROUP AND EXTRACTING WITHIN-AND BETWEEN-INDIVIDUAL VARIANCES AFTER A Z TRANSFORMATION
3) CALCULATING DIFFERENCES BETWEEN REPEATABILITIES
4) FINAL MIXED MODEL USING LME FUNCTION
5) CALCULATING VARIANCE DIFFERENCES BETWEEN CONTROL AND TREATMENT

```{r}
# #Check Normality with mixed model and then histogram of residuals
# 
# Model_Learning <- lmer(difference ~ Tank + sex + (1 | fishID), data = p2cs) 
# tab_model(Model_Learning)
# hist(residuals(Model_Learning)) 
# 
# learning_f1 <- ggplot(p2cs, aes(x=Tank, y=difference, fill = Tank)) + 
#   geom_violin()+
#   geom_jitter(height = 0, width = 0.1)+
#   facet_grid(~sex)
# 
# #Contrasts 
# em_tank <- emmeans(Model_Learning, ~ Tank) 
# pairs(em_tank)
# 
# Model_Learning_p5 <- lmer(difference ~ Tank + sex + (1 | fishID), data = p5cs) 
# tab_model(Model_Learning_p5)

#Calculating repeatabilities using custom made functions
# rpt_learning(Learning_Control) -> rpt_learning_control
rpt_learning_control
# rpt_learning(Learning_Treatment) -> rpt_learning_treatment
rpt_learning_treatment

# rpt_learning(Learning_Maternal) -> rpt_learning_Maternal
rpt_learning_Maternal
# rpt_learning(Learning_Paternal) -> rpt_learning_Paternal
rpt_learning_Paternal

#Checking combinations
groups <- c("maternal", "paternal", "control", "treatment")

combn(groups, 2)

#      [,1]       [,2]       [,3]        [,4]       [,5]        [,6]       
# [1,] "maternal" "maternal" "maternal"  "paternal" "paternal"  "control"  
# [2,] "paternal" "control"  "treatment" "control"  "treatment" "treatment"

control_boot <- unlist_rptr(rpt_learning_control) #unlisting from the control tanks
treatment_boot <- unlist_rptr(rpt_learning_treatment)#unlisting from treatment tanks
maternal_boot <- unlist_rptr(rpt_learning_Maternal)#unlisting from maternal tanks
paternal_boot <- unlist_rptr(rpt_learning_Paternal)#unlisting from paternal tanks

#Always run these again in relation to data due to same names (will fix this later)
mat_pat_diff <- difference_boot(maternal_boot,paternal_boot) #calculating the difference
qmp <- quantiles_diff_boot(mat_pat_diff) #obtaining 95% CI
mmp <- mean(mat_pat_diff) #Obtaining mean
qmp
mmp #not significant

mat_con_diff <- difference_boot(control_boot,maternal_boot) #calculating the difference
qmc <- quantiles_diff_boot(mat_con_diff) #obtaining 95% CI
mmc <- mean(mat_con_diff) #Obtaining mean
qmc
mmc #not significant


mat_tre_diff <- difference_boot(treatment_boot,maternal_boot) #calculating the difference
qmt <- quantiles_diff_boot(mat_tre_diff) #obtaining 95% CI
mmt <- mean(mat_tre_diff) #Obtaining mean
qmt
mmt #not significant

pat_con_diff <- difference_boot(control_boot,paternal_boot) #calculating the difference
qpc <- quantiles_diff_boot(pat_con_diff) #obtaining 95% CI
mpc <- mean(pat_con_diff) #Obtaining mean
qpc
mpc #SIGNIFICANT

pat_tre_diff <- difference_boot(treatment_boot,paternal_boot) #calculating the difference
qpt <- quantiles_diff_boot(pat_tre_diff) #obtaining 95% CI
mpt <- mean(pat_tre_diff) #Obtaining mean
qpt
mpt #not significant

con_tre_diff <- difference_boot(control_boot,treatment_boot) #calculating the difference
qct <- quantiles_diff_boot(con_tre_diff) #obtaining 95% CI
mct <- mean(con_tre_diff) #Obtaining mean
qct
mct #not significant

# Calculating difference in variance between control and treatment groups
Model_Learning1 <- lme(difference ~ Tank + sex, random = ~ 1| fishID, data = p2cs, na.action=na.exclude) #mixed model without variance structure
summary(Model_Learning1)
tab_model(Model_Learning1)
em_tank <- emmeans(Model_Learning1, ~ Tank) 
pairs(em_tank)
Model_Learning2 <- lme(difference ~ Tank + sex, random = ~ 1| fishID, weights = varIdent(form=~1|Tank), data = p2cs, na.action=na.exclude) #mixed model with variance structure varIdent
summary(Model_Learning2)
anova(Model_Learning1, Model_Learning2) #calculating difference between two models to find difference in variance

#Quick Visualization
p2cs %>%
  ggplot(aes(Tank, difference, fill=Tank)) +
  geom_violin(
    trim = FALSE,
    draw_quantiles = c(0.25, 0.5, 0.75), 
    alpha=0.5
  ) + 
  geom_point(position = "jitter", alpha = 0.7, size = 3)+
  facet_grid(~sex)
```
