==> Scripts/6.2_LEARNING_PLOTS.Rmd <==
---
title: "visualisation"
output: html_document
---

### Load packages
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(forcats)

# Load custom functions
source("../Scripts/functions_plots.R") #load custom functions
```

#Creating tibbles to use in plotting repeatabilities



Learning analysis
```{r, echo=FALSE}
Learning_rptr_F1 <- tibble(
  Group = c("F1C", "F1T", "F1M", "F1P"),
  r = c(rpt_learning_control$R$fishID,rpt_learning_treatment$R$fishID,rpt_learning_Maternal$R$fishID,rpt_learning_Paternal$R$fishID),
  ci.lb = c(rpt_learning_control$CI_emp$`2.5%`,rpt_learning_treatment$CI_emp$`2.5%`,rpt_learning_Maternal$CI_emp$`2.5%`,rpt_learning_Paternal$CI_emp$`2.5%`),
  ci.ub = c(rpt_learning_control$CI_emp$`97.5%`,rpt_learning_treatment$CI_emp$`97.5%`,rpt_learning_Maternal$CI_emp$`97.5%`,rpt_learning_Paternal$CI_emp$`97.5%`),
Endpoint = c("Difference", "Difference", "Difference", "Difference"))
  
Learning_rptr_F1

Learning_rptr_F1$r <- as.numeric(Learning_rptr_F1$r)


# Learning_within_between_F1 <- tibble(
#   Group = c("F1C","F1C","F1T","F1T","F1M","F1M","F1P","F1P"),
#   Condition = c("Between-Individual Variance", "Within-Individual Variance","Between-Individual Variance", "Within-Individual Variance","Between-Individual Variance", "Within-Individual Variance","Between-Individual Variance", "Within-Individual Variance"),
#   r= c(rpt_within_between_control$R[1],rpt_within_between_control$R[2],rpt_within_between_treatment$R[1],rpt_within_between_treatment$R[2],rpt_within_between_maternal$R[1],rpt_within_between_maternal$R[2],rpt_within_between_paternal$R[1],rpt_within_between_paternal$R[2]),
#   ci.lb = c(rpt_within_between_control$CI_emp[1,1],rpt_within_between_control$CI_emp[2,1],rpt_within_between_treatment$CI_emp[1,1],rpt_within_between_treatment$CI_emp[2,1],rpt_within_between_maternal$CI_emp[1,1],rpt_within_between_maternal$CI_emp[2,1],rpt_within_between_paternal$CI_emp[1,1],rpt_within_between_paternal$CI_emp[2,1]),
#   ci.ub = c(rpt_within_between_control$CI_emp[1,2],rpt_within_between_control$CI_emp[2,2],rpt_within_between_treatment$CI_emp[1,2],rpt_within_between_treatment$CI_emp[2,2],rpt_within_between_maternal$CI_emp[1,2],rpt_within_between_maternal$CI_emp[2,2],rpt_within_between_paternal$CI_emp[1,2],rpt_within_between_paternal$CI_emp[2,2]),
#   Endpoint = c("Learning","Learning","Learning","Learning","Learning","Learning","Learning","Learning")
# )
# 
# Learning_within_between_F1$r <- as.numeric(Learning_within_between_F1$r)
# 
# Learning_within_between_F1



```

#Overall forest plot showing repeatabilities of control and treatment groups (Learning F1)
```{r}
Learning_rptr_F1_reordered <- Learning_rptr_F1 %>%
  mutate(Group = fct_relevel(Group, 
            "F1T", "F1P", "F1M", 
            "F1C"))

Overall_plot_F1_learning <- ggplot(Learning_rptr_F1_reordered, aes(x=Endpoint, y=r, colour=Group)) +
geom_errorbar(aes(ymin=ci.lb, ymax=ci.ub), width = 0.4, position = position_dodge(0.3), size=0.8) +
geom_point(aes(x = Endpoint, y = r), position = position_dodge(0.3), size=3)+
geom_hline(yintercept = 0, lty = "dotted") +
  scale_colour_manual(values = c("#999999","#56B4E9","#E69F00","#009E73"), guide = guide_legend(reverse=TRUE))+
coord_flip()+
ylim(0, 0.9)+
   theme(axis.title.y=element_blank())+
  scale_linetype_manual(guide = guide_legend(reverse = TRUE))+
    theme(panel.border = element_rect(colour = "black", fill=NA, size=2),axis.ticks = element_line(size=2,color="black"),axis.ticks.length=unit(0.2,"cm"))+ font("xylab",size=15)+font("xy",size=15)+font("xy.text", size = 15) +
  theme(axis.title.y=element_blank())+
  labs(y = "ICC", title = "Repeatabiltiy of Aversive Learning")+
  scale_y_continuous(breaks=c(0,0.1,0.2,0.3,0.4,0.5,0.9))+
  theme(legend.title = element_blank())+
  annotate("text", x = 0.94, y = 0.22, label = "*", size = 7)+
  annotate("text", x = 1.10, y = 0.51, label = "*", size = 7)+
  theme(axis.text.y = element_text(angle = 90, hjust = 0.42))

Overall_plot_F1_learning

# #Within and Between individual variance plot
# Learning_within_between_F1$Group2 <- reorder(Learning_within_between_F1$Group, Learning_within_between_F1$r)
# 
# plot_within_between <- ggplot(Learning_within_between_F1, aes(x=Endpoint, y=r, colour=Group2)) +
# geom_errorbar(aes(ymin=ci.lb, ymax=ci.ub), width = 0.4, position = position_dodge(0.3), size=0.8) +
# geom_point(aes(x = Endpoint, y = r), position = position_dodge(0.3), size=3)+
# scale_colour_manual(values=c("#74D055FF", "#1F968BFF","#440154FF", "39568CFF"))+#color scheme
# geom_hline(yintercept = 0, lty = "dotted") +
# coord_flip()+
# ylim(0, 1.3)+
#    theme(axis.title.y=element_blank())+
#   theme_classic()+
#     theme(panel.border = element_rect(colour = "black", fill=NA, size=2),axis.ticks = element_line(size=2,color="black"),axis.ticks.length=unit(0.2,"cm"))+      font("xylab",size=15)+font("xy",size=15)+font("xy.text", size = 15) +
#     theme(legend.position = "bottom")+
#   theme(legend.title = element_blank())+
#   theme(axis.title.y=element_blank())+
#   facet_grid(~Condition)+
#   labs(y = "Variance component", tag = "B)")
#  
# plot_within_between
# 
# joined_plots <- Overall_plot_F1_learning + plot_within_between + plot_layout(ncol = 1, nrow=2)
# 
# joined_plots

```


#Generating custom designed violin plots to display raw distributions of behavioral responses
```{r}
groups_f1 <- c("F1C", "F1M", "F1P", "F1T")

 
p2cs$Tank2 <- reorder(p2cs$Tank, p2cs$difference)

p2cs_reordered <- p2cs %>%
  mutate(Tank = fct_relevel(Tank, 
            "Within_Control", "Maternal", "Paternal", 
            "Within_Treatment"))
 
# means <- aggregate(difference ~  Tank, p2cs, mean)

#indiv means
individual_means_learning <- p2cs_reordered %>%
    group_by(fishID, sex, Tank) %>%
    dplyr::summarize(Mean = mean(difference, na.rm=TRUE))



learning_violin1 <- violin_plot_custom_hamza(p2cs_reordered,p2cs_reordered$difference,"Difference between time spent in baseline and probe period for CS+", "Seconds")+
scale_x_discrete(labels= groups_f1)+
geom_signif(comparisons = list(c("Within_Treatment", "Paternal")), #Adding contrasts         
               map_signif_level=TRUE, y_position = 49, annotations = "*", textsize = 4)+
geom_signif(comparisons = list(c("Within_Treatment", "Maternal")), #Adding contrasts
               map_signif_level=TRUE, y_position = 42, annotations = "*", textsize = 4)  +
    stat_summary(fun=mean, geom="point", shape=18, size=3, color="red", fill="red")+
  scale_y_continuous(breaks = round(seq(min(p2cs$difference), max(p2cs$difference), by = 5),1))

learning_violin1


learning_violin2 <- violin_plot_custom_hamza(individual_means_learning,individual_means_learning$Mean,"Difference between time spent in baseline and probe period for CS+", "Seconds")+
scale_x_discrete(labels= groups_f1)+
geom_signif(comparisons = list(c("Within_Treatment", "Paternal")), #Adding contrasts         
               map_signif_level=TRUE, y_position = 30, annotations = "*", textsize = 4)+
geom_signif(comparisons = list(c("Within_Treatment", "Maternal")), #Adding contrasts
               map_signif_level=TRUE, y_position = 26, annotations = "*", textsize = 4)  +
    stat_summary(fun=mean, geom="point", shape=18, size=3, color="red", fill="red")+
  scale_y_continuous(breaks = round(seq(min(individual_means_learning$Mean), max(individual_means_learning$Mean), by = 6),1))

learning_violin2
```
